Refactor meta.js in core assets
