+++ rdmo/core/assets/js/api/BaseApi.js
@@ -1,7 +1,7 @@
 import Cookies from 'js-cookie'
 import isUndefined from 'lodash/isUndefined'
 
-import baseUrl from '../utils/baseUrl'
+import { baseUrl } from '../utils/meta'
 
 function ApiError(statusText, status) {
   this.status = status
