--- rdmo/core/assets/js/utils/baseUrl.js
@@ -1,2 +1,0 @@
-// take the baseurl from the <head> of the django template
-export default document.querySelector('meta[name="baseurl"]').content.replace(/\/+$/, '')
