--- rdmo/core/assets/js/utils/language.js
@@ -1,2 +1,0 @@
-// take the language from the <head> of the django template
-export default document.querySelector('meta[name="language"]').content
