--- rdmo/core/assets/js/utils/staticUrl.js
@@ -1,2 +1,0 @@
-// take the staticurl from the <head> of the django template
-export default document.querySelector('meta[name="staticurl"]').content.replace(/\/+$/, '')
